rename monitoring/exporter to monitoring/victoria
